requirements.txt
cython
#git+https://github.com/cocodataset/cocoapi.git#subdirectory=PythonAPI&egg=pycocotools
pycocotools
submitit
torch>=1.5.0
torchvision>=0.6.0
git+https://github.com/cocodataset/panopticapi.git#egg=panopticapi
panoptic
scipy
termcolor
addict
yapf
timm
pandas
h5py
pyyaml
matplotlib